requirements.txt
dashscope
langchain
langchain-community
langchain-openai
faiss-cpu
sentence-transformers
python-dotenv
pandas
beautifulsoup4
lxml